# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Autenticação, instituições e usuários com múltiplos perfis.",
        "title": "basis-avalia API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/plataforma/indicadores": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "indicadores-plataforma"
                ],
                "summary": "Listar o catálogo de indicadores do INEP",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    }
                }
            },
            "post": {
                "description": "Catálogo único da instalação — vale para todas as instituições imediatamente.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "indicadores-plataforma"
                ],
                "summary": "Cadastrar indicador do INEP",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/http.IndicadorPlataformaResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    }
                }
            }
        },
        "/api/v1/plataforma/indicadores/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "indicadores-plataforma"
                ],
                "summary": "Buscar indicador do INEP por ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.IndicadorPlataformaResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "indicadores-plataforma"
                ],
                "summary": "Atualizar indicador do INEP",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.IndicadorPlataformaResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    }
                }
            },
            "delete": {
                "description": "Bloqueado enquanto qualquer instituição o referenciar em meta.",
                "tags": [
                    "indicadores-plataforma"
                ],
                "summary": "Excluir indicador do INEP",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    }
                }
            }
        },
        "/api/v1/plataforma/indicadores/{id}/situacao": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "indicadores-plataforma"
                ],
                "summary": "Inativar ou reativar indicador do INEP",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.IndicadorPlataformaResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/http.erroResposta"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.IndicadorPlataformaResponse": {
            "type": "object",
            "properties": {
                "atualizado_em": {
                    "type": "string"
                },
                "codigo": {
                    "type": "string"
                },
                "criado_em": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "escopo": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metas": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "referencia_instrumento": {
                    "type": "string"
                },
                "situacao": {
                    "type": "string"
                },
                "versao": {
                    "type": "integer"
                }
            }
        },
...
        "http.erroResposta": {
            "type": "object",
            "properties": {
                "campo": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}